+++ src/main/java/Need/TimerView.java
@@ -1,6 +1,5 @@
-package UI;
+package Need;
 
-import entities.Timertomato;
 import net.miginfocom.swing.MigLayout;
 
 import javax.swing.*;
@@ -10,37 +9,37 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 @SuppressWarnings("serial")
-public class TimerView_FixCycle extends JFrame
+public class TimerView extends JFrame
 {
     private JPanel topPanel;
     String number1;
     String number2;
     // Custom Components
-    // RGB Codes for skyBlue. Used as Background.
-    private final Color skyBlue = new Color(0, 50, 200);
+    // RGB Codes for Blue. Used as Background.
+    private final Color Blue = new Color(0, 50, 200);
 
     // Used for displaying the timer values.
-    private final Font timerStyle = new Font("MonoAlphabet", Font.BOLD, 140);
+    private final Font timerStyle = new Font("Good", Font.BOLD, 140);
 
-    // RGB codes for cardinalRed. Used as Background for buttons.
-    private final Color cardinalRed = new Color(189, 32, 49);
-
-    // Used in the buttons in the Pomodoro Timer section.
-    private final Font formBTStyles = new Font("Existence", Font.BOLD, 16);
+    // RGB codes for Red. Used as Background for buttons.
+    private final Color Red = new Color(190, 30, 50);
 
     // Used in the buttons in the Pomodoro Timer section.
-    private final Font delayLabelStyles = new Font("Existence", Font.ITALIC, 20);
+    private final Font formBTStyles = new Font("formBT", Font.BOLD, 20);
 
-    // RGB code for indiaGreen. Used as background when the timer is paused.
-    private final Color indiaGreen = new Color(19, 136, 8);
+    // Used in the buttons in the Pomodoro Timer section.
+    private final Font delayLabelStyles = new Font("delaylabel", Font.ITALIC, 20);
 
-    //private static final int ORIGINAL_COUNTDOWN_MINUTES = 25;
-    private int ORIGINAL_COUNTDOWN_MINUTES = -1;
+    // RGB code for Green. Used as background when the timer is paused.
+    private final Color Green = new Color(20, 140, 30);
+
+    private static final int ORIGINAL_COUNTDOWN_MINUTES = 25;
     private static final int ORIGINAL_COUNTDOWN_SECONDS = 0;
+
     private int[] workTimeMin = {-1};
     private int[] restTimeMin = {-1};
-    //private static final int ORIGINAL_SHORTBREAK_MINUTES = 5;
-    private int ORIGINAL_SHORTBREAK_MINUTES = -1;
+
+    private static final int ORIGINAL_SHORTBREAK_MINUTES = 5;
     private static final int ORIGINAL_SHORTBREAK_SECONDS = 0;
     private static final int ORIGINAL_LONGBREAK_MINUTES = 15;
     private static final int ORIGINAL_LONGBREAK_SECONDS = 0;
@@ -53,36 +52,41 @@ public class TimerView_FixCycle extends JFrame
     private JLabel separator;
     private JLabel secondLabel;
     private JLabel delayRemainingLabel;
+
     private JButton startPauseBT;
     private JButton stopBT;
     private JButton continueBT;
+
     private Icon startIcon;
     private Icon pauseIcon;
     private Icon stopIcon;
     private Icon skipIcon;
+
     private Timer countDown;
-    private Timer shortTimer;
-    private Timer longTimer;
+    private Timer workTimer;
+    private Timer restTimer;
     private Timer delayTimer;
+
     private int secondsRemaining;
     private int minutesRemaining;
     private int delayRemaining;
     private int roundsCompleted; // No. of Pomodoro rounds completed.
 
-    private static final String workTime = Timertomato.getWorkTime();
-    private static final String restTime = Timertomato.getRestTime();
-    private static int workTime_minutes;
-    private static int workTime_seconds;
-    private static int restTimer_minutes;
-    private static int restTimer_seconds;
+    //private static final String workTime = Timertomato.getWorkTime();
+    //static String[] work = workTime.split("\\:");
 
-    private Timer workTimer;
-    private Timer restTimer;
+    //private static final int workTime_minutes = Integer.parseInt(work[0]);
+    //private static final int workTime_seconds = Integer.parseInt(work[1]);
+
+    //private static final String restTime = Timertomato.getRestTime();
+    //static String[] rest = getRestTime().split("\\:");
+    //private static final int restTimer_minutes = Integer.parseInt(rest[0]);
+    //private static final int restTimer_seconds = Integer.parseInt(rest[1]);
 
 
-    public void TimerView_FixButton()
+    public TimerView()
     {
-        setTitle("Pomodoro Timer");
+        setTitle("Tomato Timer");
 
         // Setting up the Cloud Look and Feel of the GUI application
         try
@@ -109,7 +113,7 @@ public class TimerView_FixCycle extends JFrame
         }
 
         topPanel = new JPanel();
-        topPanel.setPreferredSize(new Dimension(900, 594));
+        topPanel.setPreferredSize(new Dimension(900, 600));
         topPanel.setLayout(new BorderLayout());
         getContentPane().add(topPanel, BorderLayout.CENTER);
         roundsCompleted = 0;
@@ -121,7 +125,7 @@ public class TimerView_FixCycle extends JFrame
     private JPanel addMainTimer()
     {
         timerPanel = new JPanel();
-        timerPanel.setBackground(skyBlue);
+        timerPanel.setBackground(Blue);
         timerPanel.setLayout(new MigLayout("insets 115 0 0 0", "", "[][]0[]"));
 
         minuteLabel = new JLabel(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
@@ -139,22 +143,21 @@ public class TimerView_FixCycle extends JFrame
         secondLabel.setFont(timerStyle);
         timerPanel.add(secondLabel, "alignx center, height 145!, wrap");
 
-        //startIcon = new ImageIcon("src/Play.png");
-        //pauseIcon = new ImageIcon("src/Pause.png");
-        //startPauseBT = new JButton("Start", startIcon);
+        // startIcon = new ImageIcon("src/Play.gif");
+        // pauseIcon = new ImageIcon("src/Pause.png");
         startPauseBT = new JButton("Start");
-        startPauseBT.setContentAreaFilled(false);
-        startPauseBT.setBackground(indiaGreen);
+        startPauseBT.setContentAreaFilled(true);
+        startPauseBT.setBackground(Green);
         startPauseBT.setActionCommand("Begin");
         startPauseBT.setForeground(Color.white);
         startPauseBT.setFont(formBTStyles);
-        timerPanel.add(startPauseBT, "gaptop 10, alignx center, split 3, spanx, pushx");
+        timerPanel.add(startPauseBT, "gaptop 5, alignx center, split 3, spanx, pushx");
 
         //skipIcon = new ImageIcon("src/Skip.png");
         //continueBT = new JButton("Skip", skipIcon);
         continueBT = new JButton("Skip");
-        continueBT.setBackground(cardinalRed);
-        continueBT.setContentAreaFilled(false);
+        continueBT.setBackground(Red);
+        continueBT.setContentAreaFilled(true);
         continueBT.setForeground(Color.white);
         continueBT.setFont(formBTStyles);
         continueBT.setVisible(false);
@@ -164,12 +167,12 @@ public class TimerView_FixCycle extends JFrame
         //stopIcon = new ImageIcon("src/Stop.png");
         //stopBT = new JButton("Stop", stopIcon);
         stopBT = new JButton("Stop");
-        stopBT.setContentAreaFilled(false);
-        stopBT.setBackground(cardinalRed);
+        stopBT.setContentAreaFilled(true);
+        stopBT.setBackground(Red);
         stopBT.setActionCommand("Stop");
         stopBT.setForeground(Color.white);
         stopBT.setFont(formBTStyles);
-        //timerPanel.add(stopBT, "gaptop 10, alignx center, split 3, spanx, pushx");
+        //timerPanel.add(stopBT, "gaptop 5, alignx center, split 3, spanx, pushx");
         timerPanel.add(stopBT);
 
         delayRemainingLabel = new JLabel("Resumes in " + TOTAL_DELAY_TIME + " seconds");
@@ -177,22 +180,25 @@ public class TimerView_FixCycle extends JFrame
         delayRemainingLabel.setVisible(false);
         delayRemainingLabel.setFont(delayLabelStyles);
         timerPanel.add(delayRemainingLabel, "alignx center");
-        JTextField text3 = new JTextField(100);
-        text3.setPreferredSize(new Dimension(1,50));
-        timerPanel.add(text3);
-        JTextField text4 = new JTextField(100);
-        text4.setPreferredSize(new Dimension(1,50));
-        timerPanel.add(text4);
-        JButton button1 = new JButton("Set Time");
-        button1.setFont(new Font("", Font.PLAIN, 50));
-        timerPanel.add(button1);
-        button1.addActionListener(new ActionListener(){
+
+        JTextField workText = new JTextField(100);
+        workText.setPreferredSize(new Dimension(1,50));
+        timerPanel.add(workText, "alignx center, split 3");
+        JTextField restText = new JTextField(100);
+        restText.setPreferredSize(new Dimension(1,50));
+        timerPanel.add(restText, "alignx center, split 3");
+        JButton SettingButton = new JButton("Setting");
+        SettingButton.setFont(new Font("", Font.PLAIN, 30));
+        timerPanel.add(SettingButton, "alignx center, split 3");
+
+
+        SettingButton.addActionListener(new ActionListener(){
 
             @Override
             public void actionPerformed(ActionEvent e) {
                 // set time
-                number1 = text3.getText();
-                number2 = text4.getText();
+                number1 = workText.getText();
+                number2 = restText.getText();
                 int i1 = Integer.parseInt(number1);
                 workTimeMin[0] = i1;
                 int i2 = Integer.parseInt(number2);
@@ -203,12 +209,13 @@ public class TimerView_FixCycle extends JFrame
             }
 
         });
+
         startPauseBT.addActionListener((ActionEvent event) -> {
             switch (event.getActionCommand()) {
                 case "Pause":
                     countDown.stop();
                     countDownPaused();
-                    button1.setVisible(false);
+                    SettingButton.setVisible(false);
                     startPauseBT.setText("Continue");
                     startPauseBT.setActionCommand("Begin");
                     break;
@@ -216,14 +223,14 @@ public class TimerView_FixCycle extends JFrame
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
-                    button1.setVisible(false);
+                    SettingButton.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                 case "Begin": // Continue
                     countDown.start();
                     stopBT.setEnabled(true);
                     //startPauseBT.setIcon(pauseIcon);
-                    button1.setVisible(false);
+                    SettingButton.setVisible(false);
                     startPauseBT.setText("Pause");
                     startPauseBT.setActionCommand("Pause");
                     break;
@@ -233,14 +240,14 @@ public class TimerView_FixCycle extends JFrame
         });
 
         continueBT.addActionListener((ActionEvent event) -> {
-            if(event.getActionCommand().equals("SkipShortTimer"))
+            if(event.getActionCommand().equals("SkipRestTimer"))
             {
-                shortTimer.stop();
+                restTimer.stop();
                 roundsCompleted++;
             }
-            else if(event.getActionCommand().equals("SkipLongTimer"))
+            else if(event.getActionCommand().equals("SkipWorkTimer"))
             {
-                longTimer.stop();
+                workTimer.stop();
                 roundsCompleted = 0; // Timer is Reset.
             }
 
@@ -252,24 +259,21 @@ public class TimerView_FixCycle extends JFrame
 
         stopBT.addActionListener((ActionEvent event) -> {
             countDown.stop();
-            shortTimer.stop();
             secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
-            //minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+            // minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
             minutesRemaining = workTimeMin[0];
-            //minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+            // minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
             minuteLabel.setText(String.format("%02d", workTimeMin[0]));
             secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
             stopBT.setEnabled(false);
             roundsCompleted = 0; // Timer is Reset.
             //startPauseBT.setIcon(startIcon);
-            button1.setVisible(true);
+            SettingButton.setVisible(true);
             startPauseBT.setText("Start");
             startPauseBT.setActionCommand("Start");
         });
 
-
         runMainTimer();
-
 
         return timerPanel;
     }
@@ -277,25 +281,25 @@ public class TimerView_FixCycle extends JFrame
     private void runMainTimer()
     {
         //System.out.println("Start Main " + String.format("%d", roundsCompleted));
-        //minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
+        // minutesRemaining = ORIGINAL_COUNTDOWN_MINUTES;
         minutesRemaining = workTimeMin[0];
         secondsRemaining = ORIGINAL_COUNTDOWN_SECONDS;
 
-        //minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
+        // minuteLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_MINUTES));
         minuteLabel.setText(String.format("%02d", workTimeMin[0]));
         secondLabel.setText(String.format("%02d", ORIGINAL_COUNTDOWN_SECONDS));
 
         if(roundsCompleted == ONE_POMODORO_CYCLE)
         {
             //System.out.println("Stop Long " + roundsCompleted);
-            longTimer.stop();
+            workTimer.stop();
             roundsCompleted = 0; // Timer is Reset.
         }
 
         else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
         {
             //System.out.println("Stop Short " + roundsCompleted);
-            shortTimer.stop();
+            restTimer.stop();
         }
         // else
         {
@@ -303,11 +307,24 @@ public class TimerView_FixCycle extends JFrame
         }
 
         countDown = new Timer(INTERVAL, (ActionEvent event) -> {
-            if(secondsRemaining == 0)
+            if(secondsRemaining == ORIGINAL_COUNTDOWN_SECONDS)
             {
                 if(minutesRemaining == 0)
                 {
-                    runShortTimer();
+                    //startPauseBT.setText("Begin");
+                    //startPauseBT.setActionCommand("Start");
+                    //mainCompleted = false;
+                    //roundsCompleted++;
+
+                    // Selection of which break timer to run.
+                    //if(roundsCompleted == ONE_POMODORO_CYCLE)
+                    //{
+                       // runWorkTimer();
+                    //}
+                    //else if(roundsCompleted > 0 && roundsCompleted % 2 == 0)
+                    //{
+                        runRestTimer();
+                    //}
                 }
                 else
                 {
@@ -319,6 +336,7 @@ public class TimerView_FixCycle extends JFrame
             }
             else
             {
+                //if(secondsRemaining > ORIGINAL_COUNTDOWN_SECONDS)
                 if(secondsRemaining > 0)
                 {
                     secondsRemaining -= 1;
@@ -330,7 +348,7 @@ public class TimerView_FixCycle extends JFrame
         countDown.stop();
     }
 
-    private void runShortTimer()
+    private void runRestTimer()
     {
         //System.out.println("Start Short " + String.format("%d", roundsCompleted));
         //minutesRemaining = ORIGINAL_SHORTBREAK_MINUTES;
@@ -346,7 +364,7 @@ public class TimerView_FixCycle extends JFrame
         minuteLabel.setText(String.format("%02d", restTimeMin[0]));
         secondLabel.setText(String.format("%02d", ORIGINAL_SHORTBREAK_SECONDS));
 
-        shortTimer = new Timer(INTERVAL, (ActionEvent event) -> {
+        restTimer = new Timer(INTERVAL, (ActionEvent event) -> {
             if(secondsRemaining == ORIGINAL_SHORTBREAK_SECONDS)
             {
                 if(minutesRemaining == 0)
@@ -372,24 +390,24 @@ public class TimerView_FixCycle extends JFrame
             }
         });
 
-        shortTimer.start();
+        restTimer.start();
     }
 
-    private void runLongTimer()
+    private void runWorkTimer()
     {
         //System.out.println("Start Long " + String.format("%d", roundsCompleted));
         minutesRemaining = ORIGINAL_LONGBREAK_MINUTES;
         secondsRemaining = ORIGINAL_LONGBREAK_SECONDS;
-        stopBT.setVisible(false);
-        startPauseBT.setVisible(false);
+        stopBT.setVisible(true);
+        startPauseBT.setVisible(true);
         continueBT.setVisible(true);
-        continueBT.setActionCommand("SkipLongTimer");
+        continueBT.setActionCommand("SkipWorkTimer");
         countDown.stop();
 
         minuteLabel.setText(String.format("%02d", ORIGINAL_LONGBREAK_MINUTES));
         secondLabel.setText(String.format("%02d", ORIGINAL_LONGBREAK_SECONDS));
 
-        longTimer = new Timer(INTERVAL, (ActionEvent event) -> {
+        workTimer = new Timer(INTERVAL, (ActionEvent event) -> {
             if(secondsRemaining == 0)
             {
                 if(minutesRemaining == 0)
@@ -415,7 +433,7 @@ public class TimerView_FixCycle extends JFrame
             }
         });
 
-        longTimer.start();
+        workTimer.start();
     }
 
     private void countDownPaused()
@@ -461,7 +479,7 @@ public class TimerView_FixCycle extends JFrame
 
     public static void main(String[] args)
     {
-        SwingUtilities.invokeLater(TimerView_FixCycle::runGUI);
+        SwingUtilities.invokeLater(TimerView::runGUI);
     }
